gencrud/requirements.txt
Django==4.0.4
django-ckeditor==6.3.2
django-mptt==0.13.4
Pillow==9.1.0
transliterate                       # ?
gunicorn==20.1.0
pytz==2022.1                        # ?
docutils==0.18                      # ?
psycopg2==2.9.3
psycopg2-binary==2.9.3
easy-thumbnails==2.8.1
django-image-cropping==1.7.0        # ?
libsass==0.21.0                     # ?
django-filebrowser==4.0.1
django-filebrowser-no-grappelli
# sorl-thumbnail
django-simple-search==1.0.2
django-uuslug
django-htmlmin==0.11.0              # ?
django-daterange-filter==1.3.0
django-markdown-deux==1.0.5         # ?
django-pagedown==2.2.1              # ?
django-simple-math-captcha==2.0     # ?
django-compressor==3.1              # ?
django-import-export==2.8.0

djangorestframework==3.13.1
django-filter==21.1
markdown
